api/src/packages/authentication/authenticator.py
import bcrypt
import sqlite3

class Authentication:
    def __init__(self):
        self.pepper = 'this_is_a_secret_pepper_value'
        self.conn = sqlite3.connect("users.db")
        self.conn.execute("CREATE TABLE IF NOT EXISTS USERS (USERNAME TEXT PRIMARY KEY NOT NULL, PASSWORD TEXT NOT NULL);")
        self.conn.execute("CREATE TABLE IF NOT EXISTS PEPPER (USERNAME TEXT PRIMARY KEY NOT NULL, PEPPER TEXT NOT NULL);")
        self.conn.commit()
        self.conn.close()

    def create_user(self, username, password):
        salt = bcrypt.gensalt()
        password = password.encode('utf-8')
        password = bcrypt.hashpw(password, salt)
        conn = sqlite3.connect("users.db")
        conn.execute("INSERT INTO USERS (USERNAME, PASSWORD) VALUES (?, ?)", (username, password))
        conn.commit()
        conn.close()
        
        pepper = bcrypt.gensalt()
        conn = sqlite3.connect("users.db")
        conn.execute("INSERT INTO PEPPER (USERNAME, PEPPER) VALUES (?, ?)", (username, pepper))
        conn.commit()
        conn.close()
        
    def authenticate(self, username, password):
        conn = sqlite3.connect("users.db")
        cursor = conn.execute("SELECT PASSWORD, PEPPER FROM USERS, PEPPER WHERE USERS.USERNAME = PEPPER.USERNAME AND USERS.USERNAME =?", (username,))
        data = cursor.fetchone()
        conn.close()
        if data:
            password = password.encode('utf-8')
            password = bcrypt.hashpw(password, self.pepper)
            password = bcrypt.hashpw(password, data[1])
            return password == data[0]
        else:
            return False

    def update_password(self, username, new_password):
        salt = bcrypt.gensalt()
        new_password = new_password.encode('utf-8')
        new_password = bcrypt.hashpw(new_password, salt)
        conn = sqlite3.connect("users.db")
        conn.execute("UPDATE USERS SET PASSWORD = ? WHERE USERNAME = ?", (new_password, username))
        conn.commit()
        conn.close()








        import functools
from flask import request, jsonify

class Authentication:
    def __init__(self, auth_service):
        self.auth_service = auth_service
    
    def require_authentication(self, view_function):
        @functools.wraps(view_function)
        def wrapped_view(*args, **kwargs):
            if request.headers.get('Authorization') is None:
                return jsonify(error='Missing Authorization header'), 401
            auth_header = request.headers.get('Authorization').split()
            if auth_header[0] != 'Bearer':
                return jsonify(error='Invalid Authorization header'), 401
            username, password = auth_header[1].split(':')
            if not self.auth_service.authenticate(username, password):
                return jsonify(error='Invalid credentials'), 401
            return view_function(*args, **kwargs)
        return wrapped_view
This class takes an auth_service argument, which should be an instance of the authentication service that you created previously that has the methods for checking the credentials of the user. It has one method require_authentication which takes a view function and returns a new view function that performs authentication before calling the original view function. The new view function expects the request to contain an "Authorization" header with a value in the format Bearer {username}:{password}. It uses the passed authentication service to validate the credentials. If the authentication fails, it returns a JSON response with the error message "Invalid credentials" and HTTP status code 401 (Unauthorized).

You can use this authentication decorator by creating an instance of the class and passing an instance of your authentication service

Copy code
auth = Authentication(YourAuthService())

@app.route('/dashboard')
@auth.require_authentication
def dashboard():
    return render_template('dashboard.html')
This way, any client that wants to access the dashboard endpoint would need to include the Authorization header in the format Bearer {username}:{password} in the request.

Please note that while this example is a simple starting point, it still lacks security considerations that can compromise the security of the app.
It's important to remember that user authentication and authorization is a complex field, and should be implemented with care and professional advice if needed.