api/requirements.txt
aniso8601==9.0.1
async-timeout==4.0.2
attrs==22.2.0
cachelib==0.10.0
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
decorator==5.1.1
Flask==2.2.2
Flask-Caching==2.0.1
Flask-Cors==3.0.10
Flask-RESTful==0.3.9
flask-restx==1.0.3
gssapi==1.8.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.17.3
MarkupSafe==2.1.1
mysql-connector==2.2.9
mysql-connector-python==8.0.31
numpy==1.24.1
pandas==1.5.2
protobuf==3.20.1
pyrsistent==0.19.3
python-dateutil==2.8.2
pytz==2022.7
redis==4.4.2
requests==2.28.1
six==1.16.0
urllib3==1.26.14
Werkzeug==2.2.2
